style: structuring style folders
